+++ Tasks/Classes/ViewControllers/SummaryViewController.m
@@ -42,7 +42,7 @@
     StatsModel* statsModel = [StatsModel sharedInstance];
     
     nameLabel.text = [currentUser objectForKey:LOGGED_USER_NAME_KEY];
-    statsLabel.text = [NSString stringWithFormat:@"%d Done today, %d this\nweek. Hit rate %.1f%%",statsModel.todayCompleted, statsModel.thisWeekCompleted, [statsModel thisWeekHitRate]];
+    statsLabel.text = [NSString stringWithFormat:@"%ld Done today, %ld this\nweek. Hit rate %.1f%%",(long)statsModel.todayCompleted, (long)statsModel.thisWeekCompleted, [statsModel thisWeekHitRate]];
     UIImage* image = [EGOFileManager getImageFromPath:[currentUser objectForKey:LOGGED_USER_IMAGE_KEY]];
     
     if (image) {
